Implement mark storage.
There's a wrinkle to this: marks are exported right at the end of the
git fast-import process, and we can't know synchronously if that has
happened if we treat `git fast-import` as a dumb pipe. Therefore control
of the `git` process has now moved into `output::Output`, which was
probably going to have to happen eventually anyway.